moved login to index page, translated login errors
