+++ src/main/webapp/app/account/activate/activate.component.html
@@ -5,7 +5,7 @@
 
             <div class="alert alert-success" *ngIf="success">
                 <span jhiTranslate="activate.messages.success"><strong>Your user account has been activated.</strong> Please </span>
-                <a class="alert-link" (click)="login()" jhiTranslate="global.messages.info.authenticated.link">sign in</a>.
+                <a class="alert-link" [routerLink]="['/']" jhiTranslate="global.messages.info.authenticated.link">sign in</a>.
             </div>
 
             <div class="alert alert-danger" *ngIf="error" jhiTranslate="activate.messages.error">
