+++ src/main/webapp/app/account/activate/activate.component.ts
@@ -29,8 +29,4 @@ export class ActivateComponent implements OnInit {
             );
         });
     }
-
-    login() {
-        this.router.navigate(['login']);
-    }
 }
