+++ src/main/webapp/app/account/password-reset/finish/password-reset-finish.component.html
@@ -17,7 +17,7 @@
 
             <p class="alert alert-success" *ngIf="success">
                 <span jhiTranslate="reset.finish.messages.success"><strong>Your password has been reset.</strong> Please </span>
-                <a class="alert-link" (click)="login()" jhiTranslate="global.messages.info.authenticated.link">sign in</a>.
+                <a class="alert-link" [routerLink]="['/']" jhiTranslate="global.messages.info.authenticated.link">sign in</a>.
             </p>
 
             <div class="alert alert-danger" *ngIf="doNotMatch" jhiTranslate="global.messages.error.dontmatch">
