+++ src/main/webapp/app/account/password-reset/finish/password-reset-finish.component.ts
@@ -57,8 +57,4 @@ export class PasswordResetFinishComponent implements OnInit, AfterViewInit {
             );
         }
     }
-
-    login() {
-        this.router.navigate(['login']);
-    }
 }
