+++ src/main/webapp/app/account/register/register.component.html
@@ -117,7 +117,7 @@
             <p></p>
             <div class="alert alert-warning">
                 <span jhiTranslate="global.messages.info.authenticated.prefix">If you want to </span>
-                <a class="alert-link" (click)="openLogin()" jhiTranslate="global.messages.info.authenticated.link">sign in</a><span jhiTranslate="global.messages.info.authenticated.suffix">, you can try the default accounts:<br>- Administrator (login="admin" and password="admin") <br>- User (login="user" and password="user").</span>
+                <a class="alert-link" [routerLink]="['/']" jhiTranslate="global.messages.info.authenticated.link">sign in</a><span jhiTranslate="global.messages.info.authenticated.suffix">, you can try the default accounts:<br>- Administrator (login="admin" and password="admin") <br>- User (login="user" and password="user").</span>
             </div>
         </div>
     </div>
