+++ src/main/webapp/app/core/auth/user-route-access-service.ts
@@ -52,7 +52,7 @@ export class UserRouteAccessService implements CanActivate {
                 this.router.navigate(['accessdenied']).then(() => {
                     // only show the login dialog, if the user hasn't logged in yet
                     if (!account) {
-                        this.router.navigate(['login']);
+                        this.router.navigate(['/']);
                     }
                 });
                 return false;
