+++ src/main/webapp/app/home/home.route.ts
@@ -1,7 +1,6 @@
 import { Routes } from '@angular/router';
 
 import { HomeComponent } from './';
-import { JhiLoginComponent } from 'app/shared';
 
 export const HOME_ROUTES: Routes = [
     {
@@ -10,13 +9,6 @@ export const HOME_ROUTES: Routes = [
         data: {
             authorities: [],
             pageTitle: 'home.title'
-        }
-    }, {
-        path: 'login',
-        component: JhiLoginComponent,
-        data: {
-            authorities: [],
-            pageTitle: 'home.login'
         }
     }
 ];
