+++ src/main/webapp/app/shared/index.ts
@@ -4,7 +4,6 @@ export * from './alert/alert.component';
 export * from './alert/alert-error.component';
 export * from './auth/has-any-authority.directive';
 export * from './language/find-language-from-key.pipe';
-export * from './login/login.component';
 export * from './util/request-util';
 export * from './model/base-entity';
 export * from './pipes/safe-html.pipe';
